random messages to rp, set action

+++ src/main/java/net/artux/pdanetwork/models/user/dto/SimpleUserDto.java
@@ -14,7 +14,7 @@ public class SimpleUserDto {
     private String name;
     private String nickname;
     private String avatar;
-    private int pdaId;
+    private Long pdaId;
     private int xp;
     private int achievements;
     private Gang gang;
